+++ tests/test_mem_eff_attention.py
@@ -1216,8 +1216,6 @@ def test_grad_checkpointing(
         k,
         kv,
     ) = opFW_device_dtype_biasT_B_Mq_Mkv_H_K_Kv
-    if op is fmha.triton.FwOp:
-        pytest.skip("Triton Flash Attention 2 doesn't support backward pass yet")
     if op is fmha.triton_splitk.FwOp:
         pytest.skip("Triton Flash Decoding doesn't support backward pass yet")
     if op is fmha.ck.FwOp: